Refactor code, add audit and enhanced tests
Rewrite code make it a little cleaner and slightly faster. Not many
changes here.

Added audit to examine returned dataset looking for duplications in
`Key` and `DateTime` variables. The line count of each cyclone exists
in the raw dataset. However, I chose to ignore this as there could
potentially be typos that throw errors. Generally speaking, if there
are duplicate values for `Key` and `DateTime`, that should be
sufficient.

Enhanced tests for expected structure of returned datasets including
column names and classes. Also added tests to check options and test
options. Added tests to check for discrepancies in data as noted in
the test datasets and commit 37361a6.

All tests pass.

No errors, warnings or notes locally.

Bump version to 0.2.2.9000.